Broen/broen_lab_backend/requirements.txt
asgiref==3.10.0
crispy-bootstrap5==2025.6
Django==5.2.7
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-filter==25.2
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
